+++ pandas/tests/plotting/test_frame.py
@@ -3352,21 +3352,6 @@ class TestDataFramePlots(TestPlotBase):
         for legend, line in zip(result.get_legend().legendHandles, result.lines):
             assert legend.get_color() == line.get_color()
 
-    def test__check_ticks_props(self):
-        # GH 34768
-        df = pd.DataFrame({"b": [0, 1, 0], "a": [1, 2, 3]})
-        ax = _check_plot_works(df.plot, rot=30)
-        ax.yaxis.set_tick_params(rotation=30)
-        msg = "expected 0.00000 but got "
-        with pytest.raises(AssertionError, match=msg):
-            self._check_ticks_props(ax, xrot=0)
-        with pytest.raises(AssertionError, match=msg):
-            self._check_ticks_props(ax, xlabelsize=0)
-        with pytest.raises(AssertionError, match=msg):
-            self._check_ticks_props(ax, yrot=0)
-        with pytest.raises(AssertionError, match=msg):
-            self._check_ticks_props(ax, ylabelsize=0)
-
 
 def _generate_4_axes_via_gridspec():
     import matplotlib.pyplot as plt
